Fix bug where allocation results out of order
We were using a simpler algorithm where we use the same base amount
for every weighted part as the starting point for our shares, and then
distributing the remaining cents one-by-one round-robin style. This had
a bug where the base amount was rounded down to zero before accounting
for the specific weights, and all the money would be allocated via the
remainder distribution, which opens up the possibility of not respecting
the original order of the weights.

For example:

    # Expected
    iex> Money.allocate(~M[5], [3, 7])
    [%Money{amount: 2}, %Money{amount: 3}]

    # Actual - wrong order of results!
    iex> Money.allocate(~M[5], [3, 7])
    [%Money{amount: 3}, %Money{amount: 2}]

The new approach is a bit more complex but still simple overall.
We calculate the base amount for each base share individually,
respecting the weights and their order. The remainder is then calculated
by taking into account what's been allocated so far in our base shares.