Fix allocate/2 bug on a complex case
Looked at the allocation test cases from the ruby money gem to see if
there's any scenarios we can test against. They had a complex case
with a big positive and negative amount and a very complicated list of
fractional parts.

Had to introduce a helper function to test this, as our allocate/2 only
works with integer parts to keep it simple. Once testable, it revealed a
bug! On negative amounts, we were distributing the remainder incorrectly
based on how Kernel.rem/2 handles negative numbers.